backend/pyproject.toml
[tool.poetry]
name = "home-automation-backend"
version = "0.1.0"
description = "Home automation django backend"
authors = [
    "Maciej Rim <maciej.rim@gmail.com>",
    "Patrycja Bojdoł <pat.bojdol@gmail.com>",
    "Błażej Barcicki <blazej.barcicki@gmail.com>",
    "Tomasz Gawlik <endero238@gmail.com>"
]
license = "GPL-3.0"
repository = "https://gtihub.com/owocowe-piatki/home-automation.git"

packages = [
    { include="core", from="." },
]

[tool.poetry.scripts]
server = "core.scripts:runserver"
celery = "core.scripts:run_celery"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
django-webpack-loader = "^0.7.0"
requests = "^2.24.0"
docutils = "^0.16"
paho-mqtt = "^1.5.1"
djangorestframework = "^3.12.1"
graphene-django = "^2.13.0"
django-filter = "^2.4.0"
django-polymorphic = "^3.0.0"
channels = "^2.4.0"
graphene-subscriptions = "^1.0.2"
celery = "^5.0.2"
django-celery-beat = "^2.1.0"
django-redis = "^4.12.1"
django-celery-results = "^2.0.0"
django-fsm = "^2.7.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
isort = "^5.6.4"
pytest = "^6.1.1"
black = "^20.8b1"
flake8-black = "^0.2.1"
pytest-django = "^4.1.0"

[tool.black]
line-length=100

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = "tests.py test_*.py *_tests.py"
filterwarnings = "ignore::DeprecationWarning"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"